+++ src/common/Smi.Common/Options/GlobalOptions.cs
@@ -7,11 +7,11 @@ using System.Reflection;
 using System.Text;
 using Dicom;
 using FAnsi.Discovery;
-using Smi.Common.Messages;
 using Rdmp.Core.DataLoad.Engine.Checks.Checkers;
 using Rdmp.Core.Repositories;
 using Rdmp.Core.Startup;
 using ReusableLibraryCode.Annotations;
+using Smi.Common.Messages;
 using YamlDotNet.Serialization;
 using DatabaseType = FAnsi.DatabaseType;
 
@@ -186,8 +186,8 @@ namespace Smi.Common.Options
             var server = new DiscoveredServer(MappingConnectionString, MappingDatabaseType);
 
             var idx = MappingTableName.LastIndexOf('.');
-            var tableNameUnqualified = MappingTableName.Substring(idx +1);
-            
+            var tableNameUnqualified = MappingTableName.Substring(idx + 1);
+
             idx = MappingTableName.IndexOf('.');
             if (idx == -1)
                 throw new ArgumentException($"MappingTableName did not contain the database/user section:'{MappingTableName}'");
@@ -273,10 +273,9 @@ namespace Smi.Common.Options
 
         public ProducerOptions ReprocessingProducerOptions { get; set; }
 
-        public override string ToString()
-        {
-            return GlobalOptions.GenerateToString(this);
-        }
+        public TimeSpan SleepTime { get; set; }
+
+        public override string ToString() => GlobalOptions.GenerateToString(this);
     }
 
     /// <summary>
@@ -495,7 +494,7 @@ namespace Smi.Common.Options
             [UsedImplicitly]
             set { _extractRoot = value.TrimEnd('/', '\\'); }
         }
-        
+
         public override string ToString()
         {
             return GlobalOptions.GenerateToString(this);