+++ CHANGELOG.md
@@ -7,7 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
-...
+### Added
+
+- Added the `set-sleep-time-ms` control message to DicomReprocessor
 
 ### Changed